orderbook.java tweaked to use the enum declared inside the offer domain

+++ exchange/src/main/java/com/ermiyas/exchange/domain/settlement/CommissionPolicy.java
@@ -1,20 +1,19 @@
 package com.ermiyas.exchange.domain.settlement;
-import java.math.BigDecimal;
-
 import com.ermiyas.exchange.common.Money;
+import java.math.BigDecimal;
 public final class CommissionPolicy {
 
-    private final double rate; // e.g. 0.02 = 2%
+    private final BigDecimal rate; // e.g. 0.02 = 2%
 
     public CommissionPolicy(double rate) {
         if (rate < 0 || rate > 0.1) {
             throw new IllegalArgumentException("Invalid commission rate");
         }
-        this.rate = rate;
+        this.rate = BigDecimal.valueOf(rate);
     }
 
     public Money calculate(Money profit) {
-        return profit.multiply(BigDecimal.valueOf(rate));
+        return profit.multiply(rate);
     }
 }
